requirements.txt
asgiref==3.8.1
boto3==1.34.138
botocore==1.34.138
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.11
django-cors-headers==4.3.1
django-extensions==3.2.3
django-jazzmin==3.0.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
environs==11.0.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
marshmallow==3.21.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.2
shortuuid==1.0.13
simplejson==3.19.2
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==10.0.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.18
